moves the notification logic into the service class
